Rename RequestOption to RequestInput in preparation for RequestOutput
Instead of jamming every option type into a single table, I decided to
separate the two types of options: input and output. They operate
completely differently, and the only case where we need both types
together is when we are printing the options for the user. Therefore, it
made sense to separate the two types and create a RequestOutput struct.