Delete everything and start over
This is not a complete rewrite. While I am deleting everything, I will
be pulling parts of the code as I need them. This helps me stay
organized and not become overwhelmed with all of this code laying about.

This commit has a working readline implementation for the two states:
base and request. There are also empty files laid out as a framework.