Implement opt_candidates
It's getting a bit messy here. The problem I'm facing is that most
options require an argument, so tab completing will fail parsing. To
workaround this, I'm checking the last word to see if it looks like an
option, and if it does I'll remove it and try to parse again, then pass
that last word to opt_candidates.

Unfortunately I cannot canonicalize the option name via clap as it
parsed into my custom struct and I don't have access to the App
structure.